rename lib/fopen.* to lib/curl_fopen.*
